datanav_api/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Cors==5.0.0
psycopg[binary]==3.2.3
PyJWT==2.9.0
argon2-cffi==23.1.0
python-dotenv==1.0.1
pymongo==4.11.1
gunicorn==23.0.0
